Fix a couple of linting issues; move tests around
